backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
scikit-learn>=1.4.0
numpy>=1.26.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-json-logger>=2.0.7
tensorflow>=2.16.0
Pillow>=10.0.0
python-multipart>=0.0.9
joblib>=1.3.0
httpx>=0.27.0
